+++ tests/test_altamus_py.py
@@ -1,5 +1,5 @@
 import unittest
-from altamus_py.scan import EOSV2Scan, Header, PointFlags
+from altamus_py.scan import EOSV2Scan, Header, PointFlags, PCDEncoding
 from pathlib import Path
 import tempfile
 import simplejson
@@ -7,8 +7,11 @@ import os
 
 class TestFileGeneration(unittest.TestCase):
     def setUp(self) -> None:
-        self.file = Path("./tests/sample_files/batch_plant.bin")
+        self.complete_file = Path("./tests/sample_files/batch_plant.bin")
+        self.canceled_file = Path(
+            "./tests/sample_files/user_canceled_scan.bin")
         self.temp_dir = tempfile.TemporaryDirectory()
+        # self.temp_dir_path = Path("./tests/generated_files")
         self.temp_dir_path = Path(self.temp_dir.name)
         print("setting up for tests")
 
@@ -16,14 +19,15 @@ class TestFileGeneration(unittest.TestCase):
         print("Tearing down")
 
     def test_load_scan_from_bin(self):
-        scan = EOSV2Scan.from_path(self.file.absolute())
+        scan = EOSV2Scan.from_path(self.complete_file.absolute())
         self._test_sample_scan_header(scan.header)
         self._test_sample_scan_properties(scan)
         
-    def test_create_full_pcd_from_bin(self):
-        pcd_path = self.temp_dir_path / "all_points.pcd"
-        scan = EOSV2Scan.from_path(self.file.absolute())
-        scan.save_annotated_pcd_to_file(pcd_path)
+    def test_create_full_pcd_from_bin_binary_compressed(self):
+        pcd_path = self.temp_dir_path / "all_points_binary_compressed.pcd"
+        scan = EOSV2Scan.from_path(self.complete_file.absolute())
+        scan.save_annotated_pcd_to_file(
+            pcd_path, encoding=PCDEncoding.BINARY_COMPRESSED)
         
         self.assertTrue(os.path.exists(pcd_path))
         with open(pcd_path, "rb") as f:
@@ -33,9 +37,27 @@ class TestFileGeneration(unittest.TestCase):
         self.assertEqual(data[10000], 135) # randomly picked byte
         self.assertEqual(data[20000], 27) # randomly picked byte
     
+    def test_create_full_pcd_from_bin_ascii(self):
+        pcd_path = self.temp_dir_path / "all_points_ascii.pcd"
+        scan = EOSV2Scan.from_path(self.complete_file.absolute())
+        scan.save_annotated_pcd_to_file(pcd_path, encoding=PCDEncoding.ASCII)
+        self.assertTrue(os.path.exists(pcd_path))
+        with open(pcd_path, "rb") as f:
+            data = f.read()
+        self.assertEqual(len(data), 10276124)
+
+    def test_create_full_pcd_from_bin_binary(self):
+        pcd_path = self.temp_dir_path / "all_points_binary.pcd"
+        scan = EOSV2Scan.from_path(self.complete_file.absolute())
+        scan.save_annotated_pcd_to_file(pcd_path, encoding=PCDEncoding.BINARY)
+        self.assertTrue(os.path.exists(pcd_path))
+        with open(pcd_path, "rb") as f:
+            data = f.read()
+        self.assertEqual(len(data), 3438481)
+
     def test_create_filtered_pcd_from_bin(self):
         pcd_path = self.temp_dir_path / "healthy_only.pcd"
-        scan = EOSV2Scan.from_path(self.file.absolute())
+        scan = EOSV2Scan.from_path(self.complete_file.absolute())
         scan.save_annotated_pcd_to_file(pcd_path, include_error_points=False)
         
         self.assertTrue(os.path.exists(pcd_path))
@@ -49,7 +71,7 @@ class TestFileGeneration(unittest.TestCase):
     def test_create_scan_from_pcd(self):
         # create the pcd
         pcd_path = self.temp_dir_path / "test.pcd"
-        scan = EOSV2Scan.from_path(self.file.absolute())
+        scan = EOSV2Scan.from_path(self.complete_file.absolute())
         scan.save_annotated_pcd_to_file(pcd_path)
 
         # read in the pcd
@@ -58,20 +80,28 @@ class TestFileGeneration(unittest.TestCase):
         self._test_sample_scan_properties(scan)
 
     def test_create_polar_json_for_hyperion(self):
-        scan = EOSV2Scan.from_path(self.file.absolute())
+        scan = EOSV2Scan.from_path(self.complete_file.absolute())
         json_str = scan.polar_points_to_json()
         json = simplejson.loads(json_str)
         points = json["POINTS"]
         self.assertEqual(scan.points_count, len(points))
 
     def test_filtered_points(self):
-        scan = EOSV2Scan.from_path(self.file.absolute())
+        scan = EOSV2Scan.from_path(self.complete_file.absolute())
         healthy_only = scan.get_filtered_points([PointFlags.HEALTHY])
         self.assertTrue(len(healthy_only), scan.healthy_points_count)
 
         healthy_and_too_far = scan.get_filtered_points([PointFlags.HEALTHY, PointFlags.TOO_FAR])
         self.assertTrue(len(healthy_and_too_far), scan.healthy_points_count + scan.too_far_points_count)
     
+    def test_missing_points(self):
+        scan = EOSV2Scan.from_path(self.complete_file.absolute())
+        canceled_scan = EOSV2Scan.from_path(self.canceled_file.absolute())
+        self.assertEqual(scan.missing_points_count, 0)
+        self.assertEqual(canceled_scan.missing_points_count, 0)
+        self.assertEqual(canceled_scan.expected_points_count,
+                         canceled_scan.points_count)
+
     def _test_sample_scan_header(self, header: Header):
         # this test is specific to the provided sample "batch_plant.bin" file. If the file changes these might fail
         self.assertIsNotNone(header)
@@ -104,6 +134,10 @@ class TestFileGeneration(unittest.TestCase):
 
         # check some values from the overlap extraction property
         overlap, start, stop = scan.yaw_overlap_points
+        self.assertIsNotNone(start)
+        self.assertIsNotNone(stop)
+        if (start is None or stop is None):
+            raise ValueError("Stop or Start shouldn't be None")
         self.assertAlmostEqual(overlap, 20.1910327)
         self.assertEqual(start.size, 133496)
         self.assertEqual(stop.size, 132696)
@@ -111,6 +145,14 @@ class TestFileGeneration(unittest.TestCase):
         # total number of cartesian points should match the scan points
         cartesian = scan.cartesian_points
         self.assertEqual(len(cartesian), scan.points_count)
+
+    def test_user_canceled_scan(self):
+        scan = EOSV2Scan.from_path(self.canceled_file.absolute())
+        overlap = scan.yaw_overlap_points
+        missing = scan.missing_points_count
+        self.assertEqual(missing, 0)
+        self.assertEqual(overlap, (0, None, None))
+        print("Hello")
 
 if __name__ == "__main_":
     unittest.main()